+++ CHANGELOG.md
@@ -38,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Metroid: Samus Returns
 
-- Added: Option to change the final boss from Proteus Ridley to be Arachnus, Diggernaut, or the Metroid Queen.
+- Added: Option to change the final boss to be Arachnus, Diggernaut, Proteus Ridley (default), or the Metroid Queen.
 - Added: Changing the volume of the music and background ambience is now possible via cosmetic options.
 - Added: Option to have more Metroid DNA placed than required.
 - Changed: Area 6 - The Door to Chozo Seal West Intersection Terminal from Crumbling Stairwell is excluded from Door Lock Rando.